refactor: timer page 분리
